refactor(auth): decouple google login logic into custom hook
- Extract auth and firestore logic from loginCard.tsx to useGoogleLogin.ts`n- Improve separation of concerns between UI and business logic
